Move examples to separate file and fix linting errors
